Fix login, registration problems and prettier formatting
