+++ client/src/api/httpClient.js
@@ -1,9 +1,9 @@
 import axios from 'axios';
 
 const instance = axios.create({
-	headers: {
-    'Accept': 'application/json',
-    'Content-Type': 'application/json',
+  headers: {
+    Accept: 'application/json',
+    'Content-Type': 'application/json'
   }
 });
 
@@ -15,5 +15,19 @@ instance.interceptors.request.use((config) => {
   }
   return config;
 });
+
+instance.interceptors.response.use(
+  (response) => response,
+  (error) => {
+    const { response } = error;
+
+    // Duplicate user or invalid email/password
+    if (response.code === 409 || response.data['err'] === 'Invalid email/password') {
+      return Promise.resolve({ data: response.data });
+    }
+
+    return Promise.reject(error);
+  }
+);
 
 export default instance;
